fix: moveToFolderInCwd when cwd in hidden folder (#47)
When trying to move a folder in cwd the current implementation uses the full
file path. As a consequence if you're cwd is a hidden folder or a descendant
of a hidden folder then no options would display.

Proposed solution now is only search for the ignored options in the path
relative to the cwd. That way the currently ignored options are scoped by
the cwd.